+++ components/automate-deployment/pkg/majorupgradechecklist/checklist_manager_test.go
@@ -15,14 +15,14 @@ func TestCreatePostChecklistFile(t *testing.T) {
 	// remove json file
 	removeFile()
 	// Create json file
-	cl := NewCRUDChecklist("3")
+	cl := NewPostChecklistManager("3")
 	err := cl.CreatePostChecklistFile()
 	// return nil if json file is created successfully
 	assert.Equal(t, err, nil)
 }
 
 func TestReadPostChecklistByIdSuccess(t *testing.T) {
-	cl := NewCRUDChecklist("3")
+	cl := NewPostChecklistManager("3")
 	IsExecuted, err := cl.ReadPostChecklistById("pg_migrate")
 	// return nil if checklist by id found
 	assert.Equal(t, err, nil)
@@ -31,7 +31,7 @@ func TestReadPostChecklistByIdSuccess(t *testing.T) {
 }
 
 func TestReadPostChecklistSuccess(t *testing.T) {
-	cl := NewCRUDChecklist("3")
+	cl := NewPostChecklistManager("3")
 	result, err := cl.ReadPendingPostChecklistFile()
 	assert.Equal(t, err, nil)
 	//get json data as result
@@ -41,7 +41,7 @@ func TestReadPostChecklistSuccess(t *testing.T) {
 func TestReadPostChecklistByIdFailure(t *testing.T) {
 	// remove json file
 	removeFile()
-	cl := NewCRUDChecklist("3")
+	cl := NewPostChecklistManager("3")
 	IsExecuted, err := cl.ReadPostChecklistById("pg_migrate")
 	// return nil if checklist by id found
 	assert.Equal(t, err.Error(), "open /hab/svc/deployment-service/var/upgrade_metadata.json: no such file or directory")
@@ -52,7 +52,7 @@ func TestReadPostChecklistByIdFailure(t *testing.T) {
 func TestReadPostChecklistFailure(t *testing.T) {
 	// remove json file
 	removeFile()
-	cl := NewCRUDChecklist("3")
+	cl := NewPostChecklistManager("3")
 	result, err := cl.ReadPendingPostChecklistFile()
 	assert.Equal(t, err.Error(), "open /hab/svc/deployment-service/var/upgrade_metadata.json: no such file or directory")
 	//get json data as result
